% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertToNGVD.R
\name{convertToNGVD}
\alias{convertToNGVD}
\title{A raster used to convert feet NAVD88 to feet NGVD29}
\format{
a list with indicator region spatial data
}
\description{
Assembled from station-level data on the SOFIA USGS website. [ft. NAVD88] + [convertNGVD] = [ft. NGVD29]

\itemize{
  \item layer values, in feet, used to convert NAVD88 to NGVD29
  }
}
