% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decimalDate.R
\name{decimalDate}
\alias{decimalDate}
\title{decimalDate}
\usage{
decimalDate(rawData)
}
\arguments{
\item{rawData}{vector of dateTimes}
}
\value{
vector of decimal dates

#' @examples

\dontrun{
dateTime <- c('1984-02-28 13:56',
'1984-03-01 00:00',
'1986-03-01 00:00',
'1986-10-15 00:00')

decimalDate(dateTime)

dateTime <- c('1984-02-28', '1984-03-01',
'1986-03-01', '1986-10-15')

decimalDate(dateTime)

}
}
\description{
Create a decimal date or date/time from a vector. (function credit to EGRET)
}
