% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edenToRSM.R
\name{edenToRSM}
\alias{edenToRSM}
\title{edenToRSM}
\usage{
edenToRSM(
  data,
  mesh = vect(system.file("extdata/gis/COP_mesh", "mesh.shp", package = "RSM"), "mesh"),
  maskToEDEN = FALSE,
  func = mean
)
}
\arguments{
\item{data}{Input data, either SpatRaster or eden class.}

\item{mesh}{Geometry used for output. A SpatVector with mesh cells.}

\item{maskToEDEN}{Should data be masked to the EDEN domain? If FALSE the entire mesh is returned, including cells outside EDEN grid. If TRUE, mesh is masked to EDEN grid.}

\item{func}{Function to apply when multiple EDEN cells are within a mesh cell. Typically, a measure of central tendency.}
}
\value{
Depending on the class of the input data, either a SpatVector or an eden object (list with date and data elements)
}
\description{
Coarsen eden data to match RSM mesh cells. Function extracts data are mesh cell locations and returns a SpatVector object with the result.
}
\examples{

\dontrun{
edenDat <- fireHydro::getEDEN(returnType = 'terra', quarterly = TRUE)

eden_mesh <- edenToRSM(data = edenDat, maskToEDEN = TRUE)
plot(eden_mesh$data, ncol(eden_mesh$data), axes = FALSE, type = 'continuous')
}


}
