% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractCellData.R
\name{ExtractCellData}
\alias{ExtractCellData}
\title{ExtractCellData}
\usage{
ExtractCellData(
  variable,
  Dates,
  dimensions,
  varData,
  cell.list,
  cell.indices,
  digits.precision
)
}
\arguments{
\item{variable}{'ComputedHead'}

\item{Dates}{not sure}

\item{dimensions}{not sure}

\item{varData}{not sure}

\item{cell.list}{not sure}

\item{cell.indices}{not sure}

\item{digits.precision}{decimal places reported}
}
\value{
a dataframe.
Notes: See Notes in NetCDF_Extract()
}
\description{
Subset the requested data, round to digits.precision and create data.frame
}
