% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDateVector.R
\name{getDateVector}
\alias{getDateVector}
\title{getDateVector}
\usage{
getDateVector(ncdf, varName = "timestamps")
}
\arguments{
\item{ncdf}{an ncdf4 object (must already be loaded into environment using ncdf4 package)}

\item{varName}{name of timestamp variable in the ncdf4 object}
}
\value{
a POSIXlt vector of dates
}
\description{
This function generates a vector of dates from the timestamps of an input NetCDF file.
}
