% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDEM.R
\name{getDEM}
\alias{getDEM}
\title{getDEM}
\usage{
getDEM(ncdf_address, mesh)
}
\arguments{
\item{ncdf_address}{address of the netCDF file}

\item{mesh}{the spatial object to join output to. There must be a column named 'CellId' (case sensitive) in the mesh.}
}
\value{
output is a terra SpatVector polygon object with the DEM values in native units and datum.
}
\description{
Get land surface elevation data from RSM netCDF.
}
\examples{

\dontrun{
altq_addr <- "G:/data/models/COP/ALTQ/globalmonitors.nc"
copMesh  <- vect(system.file("extdata/gis/COP_mesh", "mesh.shp", package="RSM"),"mesh") # 6719 features

altq_dem <- getDEM(ncdf_address = altq_addr, mesh = copMesh)
}


}
