% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDifference.R
\name{getDifference}
\alias{getDifference}
\title{getDifference}
\usage{
getDifference(spdf1, spdf2, valueName = NULL)
}
\arguments{
\item{spdf1}{baseline. SPDFs can be dataframes or spatial*dataframes}

\item{spdf2}{alternative (positive value = alternative higher than baseline)}

\item{valueName}{name of column containing the quantity to compare. Should be the same in both SpatVector objects. If this argument is NULL, the first column is used}
}
\value{
output output is a SpatVector with the geometry of spdf1. Values are differences (spdf2 - spdf1) where positive values are higher in the alternative spdf2
}
\description{
gets the difference between two SpatVector objects.
}
\examples{

\dontrun{
}


}
