% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractEDEN.R
\name{getEDENbyROI}
\alias{getEDENbyROI}
\title{getEDENbyROI}
\usage{
getEDENbyROI(targetLocations, targetLocationNames, dateRange, needEDEN = TRUE)
}
\arguments{
\item{targetLocations}{locations to extract data from; must be class SpatialPointsDataFrame or SpatialPolygonsDataFrame}

\item{targetLocationNames}{option to specify the name of target locations (e.g., pts$gage)}

\item{dateRange}{Date range for which to download EDEN data}

\item{needEDEN}{whether EDEN data is needed. either TRUE, or the name of the list of EDEN data generated from lapply(dateRange, fireHydro::getEDEN).}
}
\value{
a dataframe formatted identically to the traceDataLong output from extractSimData()
}
\description{
This function extracts EDEN data for spatial locations of interest. This function can take a very long time depending on the number of regions-of-interest and their size, and if EDEN data is downloaded within the function call. NOTE: This is based on availability of data on the home page of EDEN. If data aren't there, they aren't accessible for this code.
}
\examples{

\dontrun{
dateList <- c(20200101:20200123)
newDat   <- getEDENbyROI(targetLocations = pts,
                         dateRange = dateList)
}

}
