% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFeatureIDs.R
\name{getFeatureIDs}
\alias{getFeatureIDs}
\title{getFeatureIDs}
\usage{
getFeatureIDs(
  shpFeature,
  mesh = vect(system.file("extdata/gis/nsrsm_v352", "nsrsm_mesh_landuse.shp", package =
    "RSM"), "nsrsm_mesh_landuse"),
  cellIDcolumn = "CellId"
)
}
\arguments{
\item{shpFeature}{an sf object with the indicator region, transect, point}

\item{mesh}{sf object with model mesh cells and cell ID data (mesh must have a column named "CellId")}

\item{cellIDcolumn}{name of column containing cell identifiers. Case insensitive.}
}
\value{
a list with three elements: a vector of cell IDs, (2) a vector indexing selected cells in the original mesh, and (3) a SpatVector object with the intersection of the mesh and the input shapefile
}
\description{
Extracts cell IDs from a model mesh for each indicator region. Currently only supports sf objects
}
