% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFTP.R
\name{getFTP}
\alias{getFTP}
\title{getFTP}
\usage{
getFTP(
 
    ftp = "ftp://ftppub.sfwmd.gov/outgoing/LOSOM/Iteration_2/PM_ECBr_NA25_AA_BB_CC/Model_Output/",
  destination = tmpDir(),
  pattern = "RSMBN/RSMBN_output.dss"
)
}
\arguments{
\item{ftp}{FTP site parent directory; should be a 'Model_Output' folder - see example.}

\item{destination}{destination directory. Output will appear in subfolders labeled with the alternative name.}

\item{pattern}{pattern that will define which files to download. Include any subfolders as part of this regex query.}
}
\value{
none Files will appear in subfolders of the `destination` directory. Subfolders will be labeled with the alternative name (based on the FTP folder structure).
}
\description{
downloads selected RSM output from SFWMD FTP site to a local folder. This code relies on the FTP site's folder structure, which may change at any point.
}
\examples{

\dontrun{
getFTP(ftp = 
  'ftp://ftppub.sfwmd.gov/outgoing/LOSOM/Iteration_2/PM_ECBr_NA25_AA_BB_CC/Model_Output/',
  destination = "G:\\\\data\\\\models\\\\delete",
  pattern = "RSMBN/RSMBN_output.dss") # don't pull more than one file at a time, for now.
}

}
