% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataExtract_new.R
\name{getModelData}
\alias{getModelData}
\title{getModelData}
\usage{
getModelData(
  data.source = "NetCDF",
  data.type = "stage",
  station.type = "IR",
  path.dss.ini = "../DSSvue_python/",
  file.dss.ini = "ExtractDss.ini",
  path.dss.py = "../DSSvue_python/",
  file.dss.py = "ExtractDss.py",
  data.path = "Y:/troy/RDATA/RSM_supplemental/inputFiles",
  model.version = "rsm",
  model.alternative = "COP/ECB19RR",
  indicator.region = "IR129",
  file.in = "globalmonitors.nc",
  variable = "ComputedHead",
  land.surface.datum = FALSE,
  rainFall = FALSE,
  datumOffset = 0,
  file.out = "ComputedHead.csv",
  out.path = "../../temp",
  outFile = tempfile(fileext = ".csv"),
  start.date = "1965-01-01",
  end.date = "2005-12-31",
  IR.cell.list.file = "../../listsDB/rsmWERPindicatorRegion.csv",
  cell.list = c(1317, 1533, 1892, 2333, 2508, 3117, 3329, 2942),
  indexOffset = TRUE,
  digits.precision = 3,
  DEBUG = FALSE
)
}
\arguments{
\item{data.source}{can be "dss" or "netcdf"}

\item{data.type}{can be one of the following: stage, flow, vector}

\item{station.type}{station type. options: cell, gauge, structure, IR, transect}

\item{path.dss.ini}{'../DSSvue_python/',}

\item{file.dss.ini}{'ExtractDss.ini',}

\item{path.dss.py}{'../DSSvue_python/',}

\item{file.dss.py}{'ExtractDss.py',}

\item{data.path}{paste(baseDir,'/models/rsm',sep=''),}

\item{model.version}{= 'vWERP',}

\item{model.alternative}{= 'WALT3RNL',}

\item{indicator.region}{= 'IR129',}

\item{file.in}{= 'globalmonitors.nc', # RSMGL_CEPP_output.dss, transect_flow.dss, etc.}

\item{variable}{= 'ComputedHead',}

\item{land.surface.datum}{= FALSE,}

\item{rainFall}{= FALSE,}

\item{datumOffset}{= 0.0,}

\item{file.out}{'ComputedHead.csv'}

\item{out.path}{tempdir()}

\item{outFile}{tempfile(fileext = ".csv")}

\item{start.date}{'1965-01-01'}

\item{end.date}{'2005-12-31'}

\item{IR.cell.list.file}{File address of model mesh cells in each indicator region (e.g., '../../listsDB/rsmWERPindicatorRegion.csv')}

\item{cell.list}{c( 1317, 1533, 1892, 2333, 2508, 3117, 3329, 2942 )}

\item{indexOffset}{Whether or not to add 1 to cell indices. Default =  TRUE}

\item{digits.precision}{decimal places reported}

\item{DEBUG}{not sure what this is for}
}
\value{
not sure.
}
\description{
An updated version of DataExtract() to work with streamlined . Extracts data. requires python. 

Wrapper to extract NetCDF or DSS data. The data.source controls whether ProcessNetCDF() or ProcessDSS() are called to retrive data. 
ProcessNetCDF() is governed by function input parameters, writes a .csv
file to out.path/file.out, and returns a data.frame.
ProcessDSS() is governed by a .ini file (path.dss.ini/file.dss.ini) writes a .csv file, returns NULL.
}
