% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRMap.R
\name{IRMap}
\alias{IRMap}
\title{A list with indicator region spatial data

\itemize{
  \item un-named1 tba
  \item un-named2 a SpatialPolygonDataFrame for the indicator regions
  \item col color used in plotting
  \item lwd line thickness used in plotting
  \item first tba
  }}
\format{a list with indicator region spatial data}
\usage{
data(IRMap)
}
\description{
IRMap Output from cluster analysis of the 100 simulations (all days) for the WCA-3A North Central indicator region.
}
