% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_rsm.R
\name{is.rsm}
\alias{is.rsm}
\title{Check whether an object is of the `rsm` class}
\usage{
is.rsm(x)
}
\arguments{
\item{x}{object to check}
}
\value{
logical
}
\description{
Checks object to see if it is an `rsm` object. An `rsm` object is a list with elements `cellIDs`, `cellsToUse`, `mesh`, `dates`, and `stage`
}
\examples{
is.rsm(month.abb) # FALSE

}
