% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_rsm.R
\name{loadRSM}
\alias{loadRSM}
\title{loadRSM}
\usage{
loadRSM(ncdf_address, variable = "ComputedHead", subtractTopo = TRUE)
}
\arguments{
\item{ncdf_address}{address of the netCDF file}

\item{variable}{name of variable to extract from the netCDF}

\item{subtractTopo}{should topography be subtracted? i.e., should water depths (`subtractTopo = TRUE`) or stages (`subtractTopo = FALSE`) be returned?}
}
\value{
a list with three elements: (1) a vector of dates in the time series, (2) stage data matrices (not explicitly geospatial data), and (3) a cell map that links data to cells in the RSM mesh
}
\description{
pulls stage data from an RSM output netCDF file
}
\examples{

\dontrun{
addr <- "G:/data/models/COP/ALTQ/globalmonitors.nc"
altq <- loadRSM(ncdf_address = "G:/data/models/COP/ALTQ/globalmonitors.nc")

WCA3_cellIDs <- getFeatureIDs(shpFeature = wca3,
  as(fireHydro::BICY_EVER_PlanningUnits,  "Spatial"),
  mesh = copMesh, cellIDcolumn = "CellId")
  
  hp.altq <- nc_apply(data = altq$data, # chead.altq,
  dates = altq$dateVec,
  yearBegin = 1, 
  yearlength = 12,
  spdf = ENP_cellIDs$mesh, returnSpatial = TRUE,
  func = hydroperiod, threshold = 0, continuous = FALSE) 
}


}
