% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_apply.R
\name{nc_apply}
\alias{nc_apply}
\title{nc_apply}
\usage{
nc_apply(
  data,
  cellIDs = "all",
  cellMap = NULL,
  dates = "all",
  returnSpatial = FALSE,
  spdf = NULL,
  yearBegin = 5,
  yearlength = 12,
  useParallel = FALSE,
  includeMean = TRUE,
  func = mean,
  ...
)
}
\arguments{
\item{data}{target object. function is applied to each row. If an `rsm` object is provided, several arguments are ignored: `cellIDs`, `dates`, `spdf`, `returnSpatial`}

\item{cellIDs}{which cellIDs to use? A numeric vector or "all"}

\item{cellMap}{if all cellIDs are not used, a cellMap (from ncvar_get(nc_cop, "cellmap") or loadRSM) should be provided here. This links cellIDs to the correct row in the netcdf. The first row of a cellMap must have the CellID.}

\item{dates}{a POSIXlt vector of dates}

\item{returnSpatial}{if TRUE, a joined sf is returned. If FALSE, a dataframe is returned.}

\item{spdf}{the spatial object to join output to}

\item{yearBegin}{first month of year}

\item{yearlength}{length of "year" (units = months)}

\item{includeMean}{if TRUE, a column is included that averages across all non-ID columns in the dataset (this is typically an annual average)}

\item{func}{function to apply to each year and each cell}
}
\value{
output is a vector with the value returned by "func" applied to each year

#' @examples

\dontrun{
### default behavior is to return data from the most recent date


}
}
\description{
applies a function to a netCDF object
}
