% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetCDF_Extract.R
\name{NetCDF_Extract}
\alias{NetCDF_Extract}
\title{NetCDF_Extract}
\usage{
NetCDF_Extract(
  data.path = "Y:/troy/RDATA/RSM_supplemental/inputFiles",
  model.version = "rsm",
  model.alternative = "COP/ECB19RR",
  indicator.region = "IR129",
  file.in = "globalmonitors.nc",
  variable = "olvector",
  out.file = tempfile(fileext = ".csv"),
  start.date = "1965-01-01",
  end.date = "2005-12-31",
  IR.cell.list.path = "../../listsDB",
  IR.cell.list.file = "rsmWERPindicatorRegion.csv",
  indexOffset = TRUE,
  DEBUG = TRUE
)
}
\arguments{
\item{data.path}{'Y:/troy/RDATA/RSM_supplemental/inputFiles'}

\item{model.version}{'rsm'}

\item{model.alternative}{'COP/ECB19RR'}

\item{indicator.region}{'IR129'}

\item{file.in}{'globalmonitors.nc'}

\item{variable}{'ComputedHead'}

\item{out.file}{paste0(tempfile(), ".csv")}

\item{start.date}{'1965-01-01'}

\item{end.date}{'2005-12-31'}

\item{IR.cell.list.path}{'../../listsDB'}

\item{IR.cell.list.file}{'rsmWERPindicatorRegion.csv'}

\item{indexOffset}{Whether or not to add 1 to cell indices. Default =  TRUE}

\item{DEBUG}{not sure what this is for}
}
\value{
a dataframe.
}
\description{
Read an RSM generated NetCDF file (globalmonitors.nc).
Extract the start date from the NetCDF 'timestamps' attribute.
Extract the 'cellmap' (RSM cell number to NetCDF index) matrix.
Extract data of the specified variable.
Return list of [[ Dates, dimensions, cellmap, varData ]]

Notes: Assumes that timesteps are days.

       The input file pathname is constructed as:
       data.path/model.version/model.alternative/'output'/file.in

       If variable is 1-D (Topography),   varData is a vector
       If variable is 2-D (ComputedHead), varData is 2-D matrix 
       If variable is 3-D (TotalVector),  varData is 3-D matrix

To do: Validate that 3-D "axis" dimensions correspond to 1 : X, 2 : Y
       Validate that indexOffset +1 provides the correct cell mapping
       Add ability to process multiple variables?
       Add ability for arbitray timesteps?
}
