% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetCDF_IR_Cell_Data_new.R
\name{NetCDF_IR_Cell_Data_new}
\alias{NetCDF_IR_Cell_Data_new}
\title{NetCDF_IR_Cell_Data_new}
\usage{
NetCDF_IR_Cell_Data_new(
  ncdfObj = NULL,
  indicator.region = "IR129",
  station.type = "IR",
  file.in = "globalmonitors.nc",
  variable = "ComputedHead",
  file.out = "ComputedHead.csv",
  out.path = tempdir(),
  outFile = tempfile(fileext = ".csv"),
  start.date = "1965-01-01",
  end.date = "2005-12-31",
  IR.cell.list.file = NULL,
  indexOffset = TRUE,
  digits.precision = 3,
  DEBUG = TRUE
)
}
\arguments{
\item{ncdfObj}{ncdf object loaded by getModelData (or elsewhere)}

\item{indicator.region}{'IR129'}

\item{station.type}{station type. options: cell, gauge, structure, IR, transect}

\item{file.in}{'globalmonitors.nc'}

\item{variable}{'ComputedHead'}

\item{file.out}{'ComputedHead.csv'}

\item{out.path}{tempdir()}

\item{outFile}{tempfile(fileext = ".csv")}

\item{start.date}{'1965-01-01'}

\item{end.date}{'2005-12-31'}

\item{IR.cell.list.file}{Full address of cell list file (e.g., 'rsmWERPindicatorRegion.csv')}

\item{indexOffset}{Whether or not to add 1 to cell indices. Default =  TRUE}

\item{digits.precision}{decimal places reported}

\item{DEBUG}{not sure what this is for}
}
\value{
a dataframe. The 2-D data frame will have columns of: Date, Cell_ID, Cell_ID,... 
Date       Cell_4516 Cell_4517 Cell_4518 Cell_4512
1965-01-01 17.85     17.85     17.85     17.84
1965-01-02 17.77     17.77     17.77     17.76

The 3-D data frame: 
Date       Cell_4516_X Cell_4517_X ... Cell_4516_Y Cell_4517_Y
1965-01-01 -0.00522    -0.00515    ... -0.008      -0.00604
1965-01-02 -0.00075     0.00176    ... -0.00485    -0.00456

Notes: See Notes in NetCDF_Extract()
}
\description{
Read an RSM generated NetCDF file (globalmonitors.nc).
Read the indicator.region list of RSM cell IDs for the model.version 
and model.alternative from a binary R dictionary in IR.cell.list.file
 (see R_utils/ReadIRList.R).
Subset the variable data from the indicator.region cells.
Write the data.frame to file.out as a .csv file and returns the data.frame.
}
