% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessDSS.R
\name{ProcessDSS}
\alias{ProcessDSS}
\title{ProcessDSS}
\usage{
ProcessDSS(
  path.dss.py = NULL,
  file.dss.py = NULL,
  path.dss.ini = NULL,
  file.dss.ini = NULL,
  start.date = "1965-01-01",
  end.date = "2005-12-31",
  DEBUG = FALSE
)
}
\arguments{
\item{path.dss.py}{NULL by default}

\item{file.dss.py}{NULL by default}

\item{path.dss.ini}{NULL by default}

\item{file.dss.ini}{NULL by default}

\item{start.date}{'1965-01-01'}

\item{end.date}{'2005-12-31'}

\item{DEBUG}{not sure what this is for}
}
\value{
not sure
}
\description{
See: src/DSSvue_python/ExtractDss.py
ExtractDss.py is a python 2.2.1 program passed to the HEC-DSSVue Java application.  It requires a .ini file specifying input parameters.
Note that this function creates a system2() call to the bash script hec-dssvue, which is a wrapper for hec-dssvue.sh
}
