% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessNetCDF_new.R
\name{ProcessNetCDF_new}
\alias{ProcessNetCDF_new}
\title{ProcessNetCDF_new}
\usage{
ProcessNetCDF_new(
  data.type = NULL,
  station.type = NULL,
  ncdfObj = NULL,
  indicator.region = NULL,
  file.in = NULL,
  variable = NULL,
  land.surface.datum = NULL,
  rainFall = NULL,
  datumOffset = NULL,
  file.out = NULL,
  out.path = NULL,
  outFile = tempfile(fileext = ".csv"),
  start.date = "1965-01-01",
  end.date = "2005-12-31",
  IR.cell.list.file = NULL,
  cell.list = NULL,
  indexOffset = NULL,
  digits.precision = 6,
  DEBUG = FALSE
)
}
\arguments{
\item{data.type}{not sure}

\item{station.type}{station type. options: cell, gauge, structure, IR, transect}

\item{ncdfObj}{ncdf object loaded by getModelData (or elsewhere)}

\item{indicator.region}{NULL}

\item{file.in}{'globalmonitors.nc', # RSMGL_CEPP_output.dss, transect_flow.dss, etc.}

\item{variable}{NULL}

\item{land.surface.datum}{NULL}

\item{rainFall}{NULL}

\item{datumOffset}{NULL}

\item{file.out}{NULL}

\item{out.path}{NULL}

\item{outFile}{tempfile(fileext = ".csv")}

\item{start.date}{'1965-01-01'}

\item{end.date}{'2005-12-31'}

\item{IR.cell.list.file}{full address of file.}

\item{cell.list}{c( 1317, 1533, 1892, 2333, 2508, 3117, 3329, 2942 )}

\item{indexOffset}{Whether or not to add 1 to cell indices. Default =  TRUE}

\item{digits.precision}{decimal places reported}

\item{DEBUG}{not sure what this is for}
}
\value{
not sure
}
\description{
Wrapper for various functions to extract NetCDF data
}
