% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_clusters.R
\name{sim_clusters}
\alias{sim_clusters}
\title{sim_clusters Output from cluster analysis of the 100 simulations (all days) for the WCA-3A North Central indicator region.}
\format{A dataframe with two columns}
\usage{
data(sim_clusters)
}
\description{
A dataframe grouping the 100 simulations into five clusters based on average water level.

\itemize{
  \item WCA-3ANorthCentral a column with the assigned cluster
  \item simulation         a column with the simulation number
  }
}
