% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracePlot.R
\name{tracePlot}
\alias{tracePlot}
\title{tracePlot}
\usage{
tracePlot(
  namesMatching = "[[:alnum:]]",
  simMeansObject = sim_means,
  title = "default"
)
}
\arguments{
\item{namesMatching}{a list of netCDF simulation data (e.g., "datList" object)}

\item{simMeansObject}{locations to extract data from; must be class SpatialPointsDataFrame or SpatialPolygonsDataFrame}

\item{title}{option to specify the name of target locations (e.g., pts$gage)}
}
\value{
a ggplot object
}
\description{
This function plots water depth (averaged across clustered simulations) for selected indicator regions
}
