% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertMesh.R
\name{convertMesh}
\alias{convertMesh}
\title{convertMesh}
\usage{
convertMesh(
  meshToMatch,
  meshToConvert,
  returnOriginalCellIDs = FALSE,
  output = "spdf"
)
}
\arguments{
\item{meshToMatch}{spatialpolygons to be returned. Must have "cellID" column.}

\item{meshToConvert}{cellIDs to be returned.  Must have "cellID" column.}

\item{returnOriginalCellIDs}{should MeshToMatch cellIDs be returned?}

\item{output}{type of output desired. can be either "spdf" or "df"}
}
\value{
a spatialpolygondataframe or dataframe with the cellIDs in meshToMatch that
}
\description{
Converts cell meshes using cellID centroids from meshToConvert and returns cellIDs from meshToMatch.
}
