% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCellIDs.R
\name{getCellIDs}
\alias{getCellIDs}
\title{getCellIDs}
\usage{
getCellIDs(
  modelMesh,
  indicatorRegions = RSM::IRMap[[2]],
  returnType = "polygons",
  modelName = NA,
  versionName = NA
)
}
\arguments{
\item{modelMesh}{a SpatVector with model mesh cells and cell ID data (or, a character vector of the file address to be imported internally via rgdal::readOGR)}

\item{indicatorRegions}{a SpatVector with indicator region data}

\item{returnType}{specifies whether centroid points or partial mesh polygons are returned.}

\item{modelName}{name of the model mesh used (e.g., rsm, COP)}

\item{versionName}{name of the model version used (e.g., WERP)}
}
\value{
a dataframe with cell IDs, the corresponding indicator region names and other data. Specifically, the following columns must be present: from the indicator region data: "INDICATOR", "NAME". Columns required in the mesh data: "CellId", "Node1", "Node2", "Node3", "Topo_avg", "landuse"
}
\description{
Extracts cell IDs from a model mesh for each indicator region. Currently only supports spatialPolygonDataframes; spatialLines and spatialPoints have not been tested.
}
