% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDifference.R
\name{getDifference}
\alias{getDifference}
\title{getDifference}
\usage{
getDifference(spdf1, spdf2, cellIDcolumn = "cellid_original")
}
\arguments{
\item{spdf1}{baseline. SPDFs can be dataframes or spatial*dataframes}

\item{spdf2}{alternative (positive value = alternative higher than baseline)}

\item{cellIDcolumn}{cellID column, case insensitive.}
}
\value{
output output is a spatial*dataframe with the geometry of spdf1. Values are differences (spdf2 - spdf1) where positive values are higher in the alternative spdf2
}
\description{
gets the difference between two spatial*dataframes
}
