% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFeatureIDs.R
\name{getFeatureIDs}
\alias{getFeatureIDs}
\title{getFeatureIDs}
\usage{
getFeatureIDs(
  shpFeature,
  mesh = rgdal::readOGR(system.file("extdata/gis/nsrsm_v352", "nsrsm_mesh_landuse.shp",
    package = "RSM"), "nsrsm_mesh_landuse")
)
}
\arguments{
\item{shpFeature}{a spatial*Dataframe for the indicator region, transect, point}

\item{mesh}{Shapefile with model mesh cells and cell ID data (mesh must have a column named "CellId")}
}
\value{
a list with two elements: a vector of cell IDs, and (2) a spatial dataframe with the intersection of the mesh and the input shapefile
}
\description{
Extracts cell IDs from a model mesh for each indicator region. Currently only supports spatialPolygonDataframes; spatialLines and spatialPoints have not been tested.
}
