% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_rsm.R
\name{loadRSM}
\alias{loadRSM}
\title{loadRSM}
\usage{
loadRSM(
  ncdf_address,
  variable = "ComputedHead",
  subtractTopo = TRUE,
  divideByArea = FALSE
)
}
\arguments{
\item{ncdf_address}{address of the netCDF file}

\item{variable}{name of variable to extract from the netCDF}

\item{subtractTopo}{should topography be subtracted? i.e., should water depths (`subtractTopo = TRUE`; units = feet relative to soil surface) or stages (`subtractTopo = FALSE`; units = feet NGVD29) be returned?}

\item{divideByArea}{In development. should quantity be divided by cell area? Useful for volumetric quantities: rainfall/ET.}
}
\value{
a list with class `rsm` containing three elements: (1) a vector of dates in the time series, (2) stage data matrices (not explicitly geospatial data), and (3) a cell map that links data to cells in the RSM mesh
}
\description{
pulls stage data from an RSM output netCDF file
}
\examples{

\dontrun{
addr <- "G:/data/models/COP/ALTQ/globalmonitors.nc"
altq <- loadRSM(ncdf_address = addr, 
                variable     = 'ComputedHead', 
                subtractTopo = TRUE)
}


}
