% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_apply.R
\name{nc_apply}
\alias{nc_apply}
\title{ncApply}
\usage{
nc_apply(
  data,
  cellIDs = "all",
  dates = "all",
  returnSpatial = FALSE,
  spdf = NULL,
  yearBegin = 5,
  yearlength = 12,
  includeMean = TRUE,
  func = mean,
  ...
)
}
\arguments{
\item{data}{target object. function is applied to each row. If an `rsm` object is provided, several arguments are ignored: `cellIDs`, `dates`, `spdf`, `returnSpatial`}

\item{cellIDs}{which cellIDs to use? A numeric vector or "all"}

\item{dates}{a POSIXlt vector of dates}

\item{returnSpatial}{if TRUE, a joined spdf is returned. If FALSE, a dataframe is returned.}

\item{spdf}{the spdf to join}

\item{yearBegin}{first month of year}

\item{yearlength}{length of "year" (units = months)}

\item{includeMean}{if TRUE, a column is included that averages across all non-ID columns in the dataset (this is typically an annual average)}

\item{func}{function to apply to each year and each cell}
}
\value{
output is a vector with the value returned by "func" applied to each year
}
\description{
applies a function to a netCDF object
}
