% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_apply.R
\name{rsm_apply}
\alias{rsm_apply}
\title{rsm_apply}
\usage{
rsm_apply(
  data,
  dates,
  cellMap,
  cellIDs,
  yearBegin = 1,
  yearLength = 12,
  spdf,
  returnSpatial = TRUE,
  func = function(x) {     hydroperiod(as.numeric(x), threshold = 0, continuous =
    FALSE) }
)
}
\arguments{
\item{data}{target object. function is applied to each row. If an `rsm` object is provided, several arguments are ignored: `cellIDs`, `dates`, `spdf`, `returnSpatial`}

\item{dates}{a POSIXlt vector of dates}

\item{cellMap}{if all cellIDs are not used, a cellMap (from ncvar_get(nc_cop, "cellmap") or loadRSM) should be provided here. This links cellIDs to the correct row in the netcdf. The first row of a cellMap must have the CellID.}

\item{cellIDs}{which cellIDs to use? A numeric vector or "all"}

\item{yearBegin}{first month of year}

\item{spdf}{the spatial object to join output to}

\item{func}{function to apply to each year and each cell}

\item{yearlength}{length of "year" (units = months)}

\item{includeMean}{if TRUE, a column is included that averages across all non-ID columns in the dataset (this is typically an annual average)}
}
\value{
output is a terra SpatVector polygon object with the values returned by "func" applied to each year and each cell
}
\description{
applies a function to RSM netcdf data
}
\examples{

\dontrun{
altq <- loadRSM(ncdf_address = "G:/data/models/COP/ALTQ/globalmonitors.nc")

copMesh <- vect(system.file("extdata/gis/COP_mesh", "mesh.shp", package="RSM"),"mesh") # 6719 features
copMesh$CellId <- as.integer(copMesh$CellId) # necessary?

hp.altq <- rsm_apply(data = altq$data,
dates = altq$dateVec,
cellMap = altq$cellMap,
cellIDs = copMesh$cellIDs,
yearBegin = 1,
yearlength = 12,
spdf = copMesh,
returnSpatial = TRUE,
useParallel = FALSE,
func = function(x) {hydroperiod(as.numeric(x), threshold = 0, continuous = FALSE)})  

}


}
